pyproject.toml
[project]
name = "readgen"
version = "0.0.0"
description = "A README generator for Python projects"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
authors = [
    {name = "Albert Liu", email = "dreamerhyde@gmail.com"}
]
dependencies = [
    "PyYAML>=6.0.1",
]

[project.urls]
Homepage = "https://github.com/dreamerhyde/readgen"
Repository = "https://github.com/dreamerhyde/readgen.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
readgen = "readgen.cli:main"